Redo metadataview
